Move visibility methods to their own module
