+++ compiler/noirc_frontend/src/hir/resolution/mod.rs
@@ -8,3 +8,4 @@
 pub mod errors;
 pub mod import;
 pub mod path_resolver;
+pub mod visibility;
